code/train/requirements.txt
torch~=1.7.1+cu110
Pillow~=9.4.0
torchvision~=0.8.2+cu110
tqdm~=4.65.0
numpy~=1.24.2
ftfy~=6.1.1
regex~=2023.3.23
timm~=0.5.4
einops~=0.6.0
boto3~=1.26.99
requests~=2.28.2
botocore~=1.29.99
pandas~=1.5.3